Added arm64 support to Download class

+++ Owncloud iOs Client/DataBase/Queries/ManageFilesDB.h
@@ -278,8 +278,7 @@
  * @idFile -> id of the file to update the etag
  * @etag -> new etag
  */
-+ (void) updateEtagOfFileDtoByid:(int) idFile andNewEtag: (long long) etag;
-
++ (void) updateEtagOfFileDtoByid:(NSInteger) idFile andNewEtag: (long long)etag;
 /*
  * Method to update the etag of a file
  * @FileDto
